Rename augsuggest->augprint
