+++ man/augmatch.pod
@@ -129,4 +129,4 @@ B<Augeas> project homepage L<http://www.augeas.net/>
 
 B<Augeas> path expressions L<https://github.com/hercules-team/augeas/wiki/Path-expressions>
 
-L<augsuggest>
+L<augprint>
