+++ tests/test-augprint/etc/squid/squid.conf
